Refactored controller files Moved ImageFrame to image_widget.py
